Fixed|libcommon: Misinterpretation of -skill 0
The vanilla DOS DOOM executables fail to fully range check the -skill
command line option. Due to the mechanics of thing spawning this means
that a -skill value of zero results in no things being spawned.

This logical "spawn no things" skill mode is now fully supported in
all games.

The save format version was bumped (now at version 13) in order to
separate the "fast" game rule bit from the skill mode byte.

+++ doomsday/plugins/hexen/src/p_acs.c
@@ -1654,7 +1654,7 @@ static int CmdGameType(void)
 
 static int CmdGameSkill(void)
 {
-    Push(gameSkill);
+    Push((int)gameSkill);
     return SCRIPT_CONTINUE;
 }
